TLS and relocation directories are now supported. I've begun testing the application with Corkami's handcrafted PEs: many unhandled corner-cases were identified.
BUGFIX: Directories with a size of 0 are taken into account by the Windows Loader. They aren't ignored anymore.
IMPROVEMENT: PE::_rva_to_offset now takes 64 bit integers as parameters.
BUGFIX: When resources RVA point outside of a section, we now try to use the RVA as a direct offset in the file instead of giving up.